requirements/production.txt
-r common.txt
boto==2.6.0
cssmin==0.1.4
django-storages==1.1.4
jsmin==2.0.2
psycopg2==2.4.5
